backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-genai>=1.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
python-dotenv>=1.0.0
redis>=5.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0